Small tweaks
Adjusted player movement so that they move are not longer dragging themselves along the floor (frictionless).

Adjusted player movement to look better with the animation, but increased the sprint speed modifier a tad to make sure the player still can move at a decent pace.

Started working on the environment itself. Adding bits in to make something that the player can explore.