Continued working on main game scence
Implemented an unkillable enemy that is rather dumb, where it just walks towards the player through any obstacle. But will instantly kill them if they touch it.

A ranged enemy that is able to spawn projectiles that will damage the player.

I also did some tweaking of animations to improve the movement of characters. Before, it seemed to be impacting how well everything could move.

Started working on the projectile used by the enemy ranged AI. However it is currently a dumb version that just heads in a direction.