General tweaks
Made improvements to the lighting and appearance of the main menu. Where I also added a particle emitter to produce leaves floating in the wind.
Also changed the skybox to a new texture.

Changed the enemy AI to be more difficult. They move faster, see further and spawn projectiles quicker than before.

Made some alterations to the Easter egg on the end screen, but I can't quite get it how I like it. Frankly though, it was meant to be a little added goof, so I'm not spending more time on it.